Move boost spirit headers to source file
We can move the boost spirit includes to the implementation file.
This change will increase modularity, and possibly helps us with c++20 clang modules.
